# swagger-ui/openapi.yaml
openapi: 3.0.3
info:
  title: 'Electricity Prices NordPool API'
  version: '1.0.0'
  description: |
    # Electricity Prices NordPool API
    
    Comprehensive API for accessing electricity price data from NordPool via Elering API.
    
    ## Features
    - **Multi-country support**: Lithuania (LT), Estonia (EE), Latvia (LV), Finland (FI)
    - **DST-aware timestamps**: Proper handling of daylight saving time transitions
    - **Historical data**: Access to price data from 2012-07-01 to present
    - **Real-time updates**: Latest clearing prices from NordPool
    - **Secure architecture**: All requests routed through frontend proxy
    
    ## Architecture
    - **Production**: All API calls go through frontend proxy at `/api/v1/*`
    - **Development**: Direct backend access available for debugging
    - **Security**: Backend services not exposed to internet
    
    ## Data Sources
    - **Primary**: Elering API (https://dashboard.elering.ee/api/nps/price)
    - **Storage**: PostgreSQL database with optimized queries
    - **Sync**: Automated every 30 minutes during NordPool hours
    
    ## Timezone Handling
    - **Market timezone**: Europe/Vilnius (CET/CEST)
    - **NordPool boundaries**: 22:00 UTC to 21:59 UTC
    - **DST conversion**: Automatic handling of daylight saving time
    
    ## Rate Limits
    - **No current limits**: API designed for public access
    - **Caching**: Database-based caching for optimal performance
...

...
    
  contact:
...
  license:
...
servers:
...
paths:
...
  /push/vapid-public:
    get:
      summary: 'Get VAPID public key for push subscription'
      tags:
        - Push
      responses:
        '200':
          description: 'Public key for client pushManager.subscribe'
        '503':
          description: 'Push not configured'
...
  /push/subscribe:
    post:
      summary: 'Register web push subscription'
      tags:
        - Push
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: 'Subscription stored'
        '400':
          description: 'Invalid subscription payload'
        '503':
          description: 'Push not configured'
    delete:
      summary: 'Deactivate web push subscription'
      tags:
        - Push
      responses:
        '200':
          description: 'Subscription deactivated'
  /push/test:
    post:
      summary: 'Send test push to own subscription'
      tags:
        - Push
      responses:
        '200':
          description: 'Notification sent'
        '404':
          description: 'Subscription not found'
...
  /admin/push/test:
    post:
      summary: 'Send test push to subscription by id (admin)'
      tags:
        - Admin
      security:
        - AdminToken: []
      responses:
        '200':
          description: 'Notification sent'
...
components:
...
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
...